Data management

+++ pages/baseclass/sync_screen.py
@@ -3,6 +3,7 @@ from kivy.uix.screenmanager import Screen
 from kivymd.uix.button import MDFlatButton, MDRaisedButton
 from kivymd.uix.dialog import MDDialog
 from utilities.screens import ScreensUtilities
+from utilities.storage import Storage
 
 
 class SyncScreen(Screen):
@@ -22,26 +23,41 @@ class SyncScreen(Screen):
 
     def delete_dialog(self):
         if not self.dialog:
-            self.dialog = MDDialog(
-                text="Are you sure you want to delete it?",
-                buttons=[
-                    MDFlatButton(
-                        text="Cancel", on_press=lambda x: self.dialog.dismiss()
-                    ),
-                    MDRaisedButton(
-                        text="Delete",
-                        md_bg_color=(1, 0, 0, 1),
-                        on_press=lambda x: self.delete(),
-                    ),
-                ],
-            )
+            if self.sync.in_sync:
+                self.dialog = MDDialog(
+                    text="You cannot delete it while it's syncing.",
+                    buttons=[
+                        MDFlatButton(
+                            text="Ok", on_press=lambda x: self.dialog.dismiss()
+                        ),
+                    ],
+                )
+            else:
+                self.dialog = MDDialog(
+                    text="Are you sure you want to delete it?",
+                    buttons=[
+                        MDFlatButton(
+                            text="Cancel", on_press=lambda x: self.dialog.dismiss()
+                        ),
+                        MDRaisedButton(
+                            text="Delete",
+                            md_bg_color=(1, 0, 0, 1),
+                            on_press=lambda x: self.delete(),
+                        ),
+                    ],
+                )
         self.dialog.open()
+        self.dialog.on_dismiss = self.dismiss_dialog
+
+    def dismiss_dialog(self):
+        self.dialog = None
+        return False
 
     def delete(self):
         if self.dialog:
             self.dialog.dismiss()
 
-        self.sync.delete()
+        Storage().remove(self.sync)
         ScreensUtilities().goTo("main", True)
 
     def sync_now(self):
@@ -73,10 +89,8 @@ class SyncScreen(Screen):
             self.sync_button.text = "Sync"
             self.sync_button.disabled = False
 
-
     def set_conflicts_list(self):
-        self.ids.rv.data = []
-
+        self.ids.rv.data=[]
         if self.sync.in_sync:
             self.ids.spiner.active = True
             self.ids.spiner.size = (dp(50), dp(50))
@@ -95,3 +109,7 @@ class SyncScreen(Screen):
                         ),
                     }
                 )
+
+    def on_pre_leave(self, *args):
+        Storage().unsubscribe_new_status(self.set_conflicts_list)
+        Storage().unsubscribe_new_detail(self.detail)
